requirements.txt
streamlit
matplotlib
psycopg2-binary